test(ESD-1384): add managed_account inputs, output, and prompts unit tests
- Split the monolithic managed_account_test.go into dedicated
  inputs/output/prompts test files matching the package convention.
- Add XML formatter coverage (table, JSON, CSV, XML all exercised,
  including nil/empty-slice safety cases).
- Cover explicit-field tracking on the update flag path and the error
  path on each prompt beyond the first.
